move kafka test utils to a more general place
