+++ corehq/apps/change_feed/tests/utils.py
@@ -1,0 +1,26 @@
+import uuid
+from django.conf import settings
+from kafka import KafkaConsumer
+from kafka.common import KafkaUnavailableError
+from nose.tools import nottest
+from corehq.util.test_utils import trap_extra_setup
+
+
+@nottest
+def get_test_kafka_consumer(topic):
+    """
+    Gets a KafkaConsumer object for the topic, or conditionally raises
+    a skip error for the test if Kafka is not available
+    """
+    with trap_extra_setup(KafkaUnavailableError):
+        return KafkaConsumer(
+            topic,
+            group_id='test-{}'.format(uuid.uuid4().hex),
+            bootstrap_servers=[settings.KAFKA_URL],
+            consumer_timeout_ms=100,
+        )
+
+
+def get_current_kafka_seq(topic):
+    consumer = get_test_kafka_consumer(topic)
+    return consumer.offsets()['fetch'].get((topic, 0), 0)
