+++ testapps/test_pillowtop/tests/test_case_pillow.py
@@ -2,13 +2,13 @@ import uuid
 from django.test import TestCase, override_settings
 from corehq.apps.change_feed import topics
 from corehq.apps.change_feed.consumer.feed import change_meta_from_kafka_message
+from corehq.apps.change_feed.tests.utils import get_test_kafka_consumer
 from corehq.apps.es import CaseES
 from corehq.form_processor.tests.utils import FormProcessorTestUtils
 from corehq.pillows.case import CasePillow, get_sql_case_to_elasticsearch_pillow
 from corehq.util.elastic import delete_es_index, ensure_index_deleted
 from corehq.util.test_utils import trap_extra_setup, create_and_save_a_case
 from elasticsearch.exceptions import ConnectionError
-from testapps.test_pillowtop.utils import get_test_kafka_consumer
 
 
 class CasePillowTest(TestCase):
