+++ testapps/test_pillowtop/tests/test_case_search_pillow.py
@@ -6,6 +6,7 @@ from corehq.apps.change_feed import topics
 from corehq.apps.change_feed.consumer.feed import \
     change_meta_from_kafka_message
 from corehq.apps.change_feed.producer import producer
+from corehq.apps.change_feed.tests.utils import get_test_kafka_consumer, get_current_kafka_seq
 from corehq.apps.es import CaseSearchES
 from corehq.apps.userreports.tests.utils import doc_to_change
 from corehq.elastic import get_es_new
@@ -17,8 +18,6 @@ from corehq.util.elastic import ensure_index_deleted
 from django.test import TestCase, override_settings
 from mock import MagicMock, patch
 from pillowtop.es_utils import initialize_index_and_mapping
-from testapps.test_pillowtop.utils import get_current_kafka_seq, \
-    get_test_kafka_consumer
 from corehq.util.test_utils import create_and_save_a_case
