+++ testapps/test_pillowtop/tests/test_ledger_pillow.py
@@ -4,6 +4,7 @@ from elasticsearch.exceptions import ConnectionError
 from casexml.apps.case.mock import CaseFactory
 from corehq.apps.change_feed import topics
 from corehq.apps.change_feed.consumer.feed import change_meta_from_kafka_message
+from corehq.apps.change_feed.tests.utils import get_test_kafka_consumer
 from corehq.elastic import get_es_new
 from corehq.form_processor.parsers.ledgers.helpers import UniqueLedgerReference
 from corehq.form_processor.tests.utils import FormProcessorTestUtils, run_with_all_backends
@@ -13,7 +14,6 @@ from corehq.pillows.mappings.ledger_mapping import LEDGER_INDEX_INFO
 from corehq.util.elastic import ensure_index_deleted
 from corehq.util.test_utils import trap_extra_setup
 from pillowtop.es_utils import initialize_index_and_mapping
-from testapps.test_pillowtop.utils import get_test_kafka_consumer
 
 
 class LedgerPillowTest(TestCase):
