+++ testapps/test_pillowtop/utils.py
@@ -1,31 +1,9 @@
-import uuid
-from kafka import KafkaConsumer
-from kafka.common import KafkaUnavailableError
-
-from corehq.util.decorators import ContextDecorator
-from corehq.util.test_utils import trap_extra_setup
-from nose.tools import nottest
-from pillowtop import get_pillow_by_name
 from django.conf import settings
 
-@nottest
-def get_test_kafka_consumer(topic):
-    """
-    Gets a KafkaConsumer object for the topic, or conditionally raises
-    a skip error for the test if Kafka is not available
-    """
-    with trap_extra_setup(KafkaUnavailableError):
-        return KafkaConsumer(
-            topic,
-            group_id='test-{}'.format(uuid.uuid4().hex),
-            bootstrap_servers=[settings.KAFKA_URL],
-            consumer_timeout_ms=100,
-        )
+from corehq.apps.change_feed.tests.utils import get_current_kafka_seq
 
-
-def get_current_kafka_seq(topic):
-    consumer = get_test_kafka_consumer(topic)
-    return consumer.offsets()['fetch'].get((topic, 0), 0)
+from corehq.util.decorators import ContextDecorator
+from pillowtop import get_pillow_by_name
 
 
 class process_kafka_changes(ContextDecorator):
